requirements.txt
flask==3.0.3
flask-sqlalchemy==3.1.1
pyyaml==6.0.2
python-slugify==8.0.1
typing-extensions==4.9.0
